+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ## 2.13 branch
 ### 2.13.5
+* the `RotateCommand` class is deprecated and will be removed in a later release. For this reason, the `ExportCommand` class now uses the `BackupManager::rotate()` method to
+  continue supporting the `--rotate` option;
 * compatibility with the transition from `_cake_core_` to `_cake_translations_` expected in CakePHP 5.1;
 * updated for the latest version of psalm.